Share early-exit embedding extraction between app and infer paths.
This moves the helpers into utils/embedding.py and updates both entrypoints to use the same implementation instead of duplicating logic in app.py only.
